src/test/java/com/panibrat/array/_977/SolutionTest.java
package com.panibrat.array._977;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.Test;

/*
977. Squares of a Sorted Array

Given an array of integers A sorted in non-decreasing order, return an array of the squares of each number, also in sorted non-decreasing order.

Example 1:
Input: [-4,-1,0,3,10]
Output: [0,1,9,16,100]

Example 2:
Input: [-7,-3,2,3,11]
Output: [4,9,9,49,121]

Note:
* 1 <= A.length <= 10000
* -10000 <= A[i] <= 10000
* A is sorted in non-decreasing order.

 */
public class SolutionTest {

  @Test
  public void sortedSquares() {
    assertThat(new Solution().sortedSquares(new int[] {-4, -1, 0, 3, 10}))
        .containsExactly(0, 1, 9, 16, 100);
    assertThat(new Solution().sortedSquares(new int[] {-7, -3, 2, 3, 11}))
        .containsExactly(4, 9, 9, 49, 121);
  }
}
